export frontpagesearch, refactor
